added function to reduce size of a class representation to a defined subset size (#27)

+++ src/preprocessing/spotify_dataset_generator.py
@@ -10,6 +10,7 @@ import sys
 import time
 
 from dask import dataframe as dd
+from pathlib import Path
 from spotipy.oauth2 import SpotifyClientCredentials
 
 # Set logging level
@@ -22,14 +23,20 @@ SPOTIPY_CLIENT_ID = os.environ.get("SPOTIPY_CLIENT_ID")
 SPOTIPY_CLIENT_SECRET = os.environ.get("SPOTIPY_CLIENT_SECRET")
 
 # Initialize target datasets
-TARGET_FILE = "data/all_classified_billboard_songs1.csv"
-TARGET_CLEANED_FILE = TARGET_FILE.split(".csv")[0] + "_clean.csv"
-TEMP_DIR = "data/temp"
+TARGET_FILE = (
+    Path(__file__).parents[2].joinpath("data/all_classified_billboard_songs1.csv")
+)
+TARGET_CLEANED_FILE = str(TARGET_FILE.resolve()).split(".csv")[0] + "_clean.csv"
+TEMP_DIR = Path(__file__).parents[2].joinpath("data/temp")
 BB_SUBSET_PREFIX = f"{TEMP_DIR}/bb_subset_clean"
 INTERMEDIATE_SUBSET_PREFIX = f"{TEMP_DIR}/bb_subset_id"
 SPOTIFY_SUBSET_PREFIX = f"{TEMP_DIR}/spotify_enhanced_dataset"
-SPOTIFY_DATASET_FILE = "data/spotify_enhanced_dataset.csv"
-MILLION_SONG_SUBSET_FILE = "data/million_songs_subset.csv"
+SPOTIFY_DATASET_FILE = (
+    Path(__file__).parents[2].joinpath("data/spotify_enhanced_dataset.csv")
+)
+MILLION_SONG_SUBSET_FILE = (
+    Path(__file__).parents[2].joinpath("data/million_songs_subset.csv")
+)
 
 RANDOM_SEED = 0
 
@@ -350,16 +357,38 @@ def combine_spotify_dataset_shards():
         sp_enhanced_df.to_csv(SPOTIFY_DATASET_FILE, index=False)
 
 
+def regularize_spotify_dataset_distribution(
+    filename: Path, labels: list[int], n: int = 3000
+) -> None:
+    """
+    Regularizes Spotify dataset distribution by undersampling majority class
+    """
+    new_filename = f"{str(filename).split('.csv')[0]}_regularized.csv"
+    if not os.path.exists(new_filename):
+        df = pd.read_csv(filename, on_bad_lines="skip")
+        for label in labels:
+            df_label = df[df["class"] == label]
+            if len(df_label) > n:
+                df_label = df_label.sample(n=n, random_state=RANDOM_SEED)
+                df = df[df["class"] != label]
+                df = pd.concat([df, df_label], axis=0, ignore_index=True)
+        df.to_csv(new_filename, index=False)
+    else:
+        logging.info(f"{new_filename} already exists.")
+
+
 def _delete_temp_files():
     """Deletes temporary files"""
-    logging.info("Deleting temporary files...")
-    shutil.rmtree(TEMP_DIR)
+    if os.path.exists(TEMP_DIR):
+        logging.info("Deleting temporary files...")
+        shutil.rmtree(TEMP_DIR)
 
 
 def main():
     create_cleaned_billboard_dataset()
     create_spotify_dataset()
     _delete_temp_files()
+    regularize_spotify_dataset_distribution(SPOTIFY_DATASET_FILE, [0, 10], 3000)
 
 
 if __name__ == "__main__":
